% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mush.R
\name{mush}
\alias{mush}
\title{fuse multiple layers}
\usage{
mush(...)
}
\arguments{
\item{...}{entered arguments of sf polygons layers}
}
\value{
sf polygon layer with attributes, a, b, c, ... as per length of '..."
}
\description{
Combine polygon layers into one partition of the plane.
}
\details{
Each layer (sf polygon object) is entered as an argument, may be 2 or more layers.
}
\examples{
library(basf)
afile <- system.file("extdata/list_parcels_hobart.gdb", package = "mush", mustWork = TRUE)
bfile <- system.file("extdata/list_tasveg_30_hobart.gdb", package = "mush", mustWork = TRUE)
A <- sf::read_sf(afile)
B <- sf::read_sf(bfile)
x <-  mush(A, B)
}
